refactor(desktop): clean up React effect ownership
